requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
google-genai>=1.0.0
python-dotenv==1.0.1
